% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_env.R
\name{create_env}
\alias{create_env}
\title{Create your conda virtual env with DALEX}
\usage{
create_env(yml, condaenv)
}
\arguments{
\item{yml}{string - a path to the .yml file. If OS is Windows conda has to be added to the PATH first}

\item{condaenv}{- a path to main conda folder. If OS is Unix You have to specify it with .yml file path. Using with windows, param will be omitted.}
}
\value{
Name of created virtual env.
}
\description{
Phython objects may be loaded into R. However, it requiers versions of the Python and libraries to match bewtween both machines.
}
\examples{

if(.Platform$OS.type=="windows"){
  create_env(system.file("extdata", "scikitlearn.yml", package = "DALEXtra"))
}else{
  print("Use windows for tests")
}

}
\author{
Szymon Maksymiuk
}
