% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dalex_load_explainer.R
\name{dalex_load_explainer}
\alias{dalex_load_explainer}
\title{DALEX load explainer}
\usage{
dalex_load_explainer(path)
}
\arguments{
\item{path}{Path to the pickle file with explainer saved.}
}
\description{
Load DALEX explainer created with Python library into the R environment.
}
\details{
Function uses the \code{reticulate} package to load Python object saved
in a pickle and make it accessible within R session. It also adds explainer
class to the object so it can be used with DALEX R functions.
}
