% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overall_comparison.R
\name{print.overall_comparison}
\alias{print.overall_comparison}
\title{Print overall_comparison object}
\usage{
\method{print}{overall_comparison}(x, ...)
}
\arguments{
\item{x}{an object of class \code{overall_comparison}}

\item{...}{other parameters}
}
\description{
Print overall_comparison object
}
\examples{
library("DALEXtra")
library("mlr")
task <- mlr::makeRegrTask(
  id = "R",
  data = apartments,
  target = "m2.price"
)
learner_lm <- mlr::makeLearner(
  "regr.lm"
)
model_lm <- mlr::train(learner_lm, task)
explainer_lm <- explain_mlr(model_lm, apartmentsTest, apartmentsTest$m2.price, label = "LM")

learner_rf <- mlr::makeLearner(
  "regr.randomForest"
)
model_rf <- mlr::train(learner_rf, task)
explainer_rf <- explain_mlr(model_rf, apartmentsTest, apartmentsTest$m2.price, label = "RF")

learner_svm <- mlr::makeLearner(
  "regr.ksvm"
)
model_svm <- mlr::train(learner_svm, task)
explainer_svm <- explain_mlr(model_svm, apartmentsTest, apartmentsTest$m2.price, label = "SVM")

data <- overall_comparison(explainer_lm, list(explainer_svm, explainer_rf), type = "regression")
plot(data)
}
