% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_test_comparison.R
\name{print.training_test_comparison}
\alias{print.training_test_comparison}
\title{Print funnel_measure object}
\usage{
\method{print}{training_test_comparison}(x, ...)
}
\arguments{
\item{x}{an object of class \code{funnel_measure}}

\item{...}{other parameters}
}
\description{
Print funnel_measure object
}
\examples{
library("DALEXtra")
library("mlr")
task <- mlr::makeRegrTask(
  id = "R",
  data = apartments,
  target = "m2.price"
)
learner_lm <- mlr::makeLearner(
  "regr.lm"
)
model_lm <- mlr::train(learner_lm, task)
explainer_lm <- explain_mlr(model_lm, apartmentsTest, apartmentsTest$m2.price, label = "LM")

learner_rf <- mlr::makeLearner(
  "regr.randomForest"
)
model_rf <- mlr::train(learner_rf, task)
explainer_rf <- explain_mlr(model_rf, apartmentsTest, apartmentsTest$m2.price, label = "RF")

learner_svm <- mlr::makeLearner(
  "regr.ksvm"
)
model_svm <- mlr::train(learner_svm, task)
explainer_svm <- explain_mlr(model_svm, apartmentsTest, apartmentsTest$m2.price, label = "SVM")

data <- training_test_comparison(explainer_scikit, list(explainer_h2o, explainer_mlr), 
                                 training_data = titanic_train[,-18], 
                                 training_y = titanic_train[,18])
print(data)
}
