% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_env.R
\name{create_env}
\alias{create_env}
\title{Create your conda virtual env with DALEX}
\usage{
create_env(yml, condaenv)
}
\arguments{
\item{yml}{a path to the .yml file. If OS is Windows conda has to be added to the PATH first}

\item{condaenv}{path to main conda folder. If OS is Unix You have to specify it with .yml file path. Using with windows, param will be omitted.}
}
\value{
Name of created virtual env.
}
\description{
Phython objects may be loaded into R. However, it requiers versions of the Python and libraries to match bewtween both machines.
}
\examples{
is_conda <- function(){
is_conda <- try(reticulate::conda_binary())
class(is_conda) != "try-error"
}
if(is_conda()) {
  reticulate::use_condaenv("myenv")

  if (!("myenv" \%in\% reticulate::conda_list()$name)) {
    create_env(
      system.file("extdata", "scikitlearn_unix.yml", package = "DALEXtra"),
      condaenv = paste(
        sub('[/][^/]+$', '', reticulate::conda_binary()[1]),
        "/..",
        sep = ""
      )
    )
  } else{
    print("env already exists")
  }
} else {
  "conda is required"
}
}
\author{
Szymon Maksymiuk
}
