% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_env.R
\name{create_env}
\alias{create_env}
\title{Create your conda virtual env with DALEX}
\usage{
create_env(yml, condaenv)
}
\arguments{
\item{yml}{a path to the .yml file. If OS is Windows conda has to be added to the PATH first}

\item{condaenv}{path to main conda folder. If OS is Unix You may want to specify it. When passed with windows, param will be omitted.}
}
\value{
Name of created virtual env.
}
\description{
Python objects may be loaded into R. However, it requiers versions of the Python and libraries to match between both machines.
This functions allow user to create conda virtual environment based on provided .yml file.
}
\examples{
\donttest{
  create_env(system.file("extdata", "testing_environment.yml", package = "DALEXtra"))
}
}
\author{
Szymon Maksymiuk
}
