% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yhat.R
\name{yhat_mlr}
\alias{yhat_mlr}
\alias{yhat_h2o}
\title{Wrapper over the predict function}
\usage{
yhat_mlr(X.model, newdata, ...)

yhat_h2o(X.model, newdata, ...)
}
\arguments{
\item{X.model}{object - a model to be explained}

\item{newdata}{data.frame or matrix - observations for prediction}

\item{...}{other parameters that will be passed to the predict function}
}
\value{
An numeric vector of predictions
}
\description{
These functios are default predict functions.
Each function returns a single numeric score for each new observation.
}
\details{
Currently supported packages are:
\itemize{
\item `mlr` see more in explain_mlr
\item `h2o` see more in explain_h2o
}
}
